Component Map has been rewritten to hoc
