+++ src/components/container/container.test.js
@@ -1,0 +1,14 @@
+import React from "react";
+import renderer from "react-test-renderer";
+import Container from "./container";
+
+it(`Render correctly component Container`, () => {
+  const component = renderer.create(
+      <Container
+        id={`id`}
+        style={{}}
+      />
+  ).toJSON();
+
+  expect(component).toMatchSnapshot();
+});
